Keep BuildRequest db entry around longer
The lifetime of the BuildRequest db entry needs to be longer than it
currently is. As long as there is an InstanceMapping entry with a null
CellMapping the BuildRequest object may be needed for API requests. This
moves the removal of that db entry until after the InstanceMapping has
been updated to point at a cell, or removed.

Change-Id: I2b44ae9dfba4066bc7c9a92644e3782f76097c21
Partially-implements: bp add-buildrequest-obj