libvirt: attach configdrive after instance XML
In order to expose device tagging metadata in the configdrive, the
latter needs to be attached after the instance XML has been generated
and device addresses are known, but before the guest OS runs, as
it may depend on cloud-init data on the config drive.

This patch moves configdrive creation after the instance XML has
been generated but before the guest runs.

Partially implements: blueprint virt-device-role-tagging
Co-authored-by: Vladik Romanovsky <vromanso@redhat.com>
Change-Id: I931421ea688641e2ceb212c6dc099639c53433f2