Fix incorrect format string
The 'id' in the FlavorExtraSpecUpdateCreateFailed message is actually
for the flavorid (not Flavor.id) which can be a string. This was
originally fixed in the following commit, but unfortunately the fix was
coupled with EngineFacade changes, so it's not a simple backport. This
patch just splits out the format string bug fix, so it can be
backported.

    enginefacade: 'flavor' - Use enginefacade in 'flavor' section.
    https://github.com/openstack/nova/commit/c85cd57c5c156b479db75c9f14fd916f7b64298e

Closes-Bug: #1528250
Change-Id: I2f57bb46d087948792e749905a0c060d009fbcf8
(cherry picked from commit c85cd57c5c156b479db75c9f14fd916f7b64298e)