increase error handling for dirty files
In working on converting data over to this new format, it will often
show up dirty with things like a broken reference to the parameters
file. We can make the parser be a bit more robust to these errors and
explain what is going on both in a warning, and in the final document.

Change-Id: I9cd44bddffadd20bfe4f3c12f2102f7e9a69c909