Make some build_requests columns nullable
Now that a serialized Instance object is being stored in the
build_requests table most of the other columns are no longer necessary
and should be deprecated. By first allowing them to be nullable they can
then be removed from the BuildRequest object and then removed in Ocata.

Also remove a UniqueConstraint that is no longer necessary.

Change-Id: I92fa2381c43e5e05d95bbcf9a7bb4306d3b3b856
Partially-implements: bp add-buildrequest-obj