virt: set address space & CPU time limits when running qemu-img
This uses the new 'prlimit' parameter for oslo.concurrency execute
method, to set an address space limit of 1GB and CPU time limit
of 2 seconds, when running qemu-img.

This is a re-implementation of the previously reverted commit

commit da217205f53f9a38a573fb151898fbbeae41021d
Author: Tristan Cacqueray <tdecacqu@redhat.com>
Date:   Wed Aug 5 17:17:04 2015 +0000

    virt: Use preexec_fn to ulimit qemu-img info call

NOTE(mriedem): The backport depends on raising the minimum
required oslo.concurrency>=3.7.1 which contains the prlimits
fix for this change, the code fails without it.

Depends-on: I157a6fc88836da73814ae46c8991ddefc5066c96
Closes-Bug: #1449062
Change-Id: I135b5242af1bfdcb0ea09a6fcda21fc03a6fbe7d
(cherry picked from commit 068d851561addfefb2b812d91dc2011077cb6e1d)