Fix reverse_upsize_quota_delta attempt to look up deleted flavors
When we did the "great flavor migration of 2015" we missed a quota method
which still looks up flavors by id from the migration. Now that flavors
are moved to the api database and actually removed when deleted, this no
longer works. The problem manifests itself as a failure when trying to
revert a migration or resize operation when the original flavor has been
deleted.

Change-Id: I5f95021410a309ac07fe9f474cbcd0214d1af208
Closes-Bug: #1570748
(cherry picked from commit a46e847aad4ee7edbb63eb08f97f6635e6c9ccb0)