network: Fix nova boot with multiple security-groups
Booting an instance with 'default' + other security-groups doesn't
apply the 'default' security-group on the instance.
Nova is required to remove the 'default' security-group from the request
to Neutron when it's the only security-group that should be applied on
the instance.
This bug is a result of change
I630008a9733624a9d9b59b7aa3b8b2a3f8985d61.

Closes-Bug: #1572171
Change-Id: I3eb685f4d82615d3ebd25e7471de47129c78c86a