Fix database poison warnings, part 5
The following warning appears in the unit test logs a number of times.
This patch fixes the last warnings in unit.compute.test_compute_api.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I9f81e22fda8dbab2007aa7d2aa972b9dce364b5f
Related-Bug: #1568414