pci: Deprecate is_new from pci requests
Patch https://review.openstack.org/#/c/307124/
stash the the old/new pci requests in the
migration context same as we do for NUMA.

This patch is a cleanup patch to deprecate the is_new field
in pci request. The is_new was horrible hack to get the
newer pci request on resize/migration and since the
pci_requests are now stashed in the migration context we
can deprecate it.

Change-Id: Ie36a74fb2aad9a14a2dc0e0b53649ee770fb2f9d