Make Keypair object favor the API database
This makes the keypair object load first from the API database,
falling back to the main database as necessary. Creates happen in the
API database only now.

Since the Instance object uses this object to patch up its keypairs
field on the fly, this adds a localonly parameter to the get method so
that we don't end up with compute nodes calling back up to the API
database in the lazy-load case. Only Instances that were created before
the previous patch will be missing keypair data, and so we're
guaranteed to have their keypairs in the main database (if at all).

Related to blueprint cells-keypairs-api-db

Change-Id: I700cf8633f694c933f17dd133fa6c84c2beac4c0