config options: centralize default flavor option
The default flavor was introduced for the ec2 API. That API is
deprecated and this option is not used in functional code anymore
but in unit tests.
A follow up patch will deprecate this option. In the next cycle we
can then remove it. I've put the option to its own module to avoid
merge conflicts with the nova.conf.compute module.
As this option was the last one for the "oslo.config" generator
namespace "nova.compute", the generator file was changed too and
the nova.compute.ops module was deleted.

blueprint centralize-config-options-newton

Change-Id: Ie2be972673498ef984dfd3b5164512d77ef85ace