Fix service version to update the DB
When new code is installed with a newer service version, the version in
the DB should get updated.

The fix here is to:
1) eliminate the special case for 'version' in save()
2) cause version to be saved on startup

NOTE(lyarwood): Single conflict caused by the use of Mock 1.3.0 on
stable/mitaka and as a result the lack of assert_called_once().

Conflicts:
        nova/tests/unit/test_service.py

Change-Id: I96fa9dabfb9b7a5f1703baf80534d8b104dab4e6
Closes-Bug: 1579839
(cherry picked from commit 854c39e26d94f006f539ea993ffaafa4892b100c)