deprecate "file transfer" feature for Glance images
In August 2013 the blueprint "image-multiple-location" got merged
with commit 6f9ed56. This commit introduced another way of
downloading images from glance, despite the normal way via HTTP.

After discussing this in IRC, we believe that this feature was never
used in any production environment (it was designed for a Rackspace
use case which never made it to reality) and that the plugin mechanism
of image downloaders isn't the way to go, this change will deprecate
the feature.

Change-Id: I67e2cbf956286132da90ad6a4c6d5075d99177a9