api-ref: finish validation for os-server-external-events.inc
This completes the validation for the admin event interface, including
detailing the codes we currently return in the parameter list.

Because os-service was incorrectly using the 'status' key, when I
renamed status => event_status I needed to also create a
service_status to make things compile. It's accurate enough to move
forward.

Because this is really a Neutron only API, this is moved towards the
bottom of the API list, to make it clear this isn't really designed to
be used by end user apps.

Part of bp:api-ref-in-rst

Change-Id: I3ed46eec1c33a63b7a89d6d6b5eeda574effdc56