report info if parameters are out of order
This puts in the infrastructure for enforcing the section and
parameter name ordering in parameters.yaml file. It does so through
loading the lookup table as an ordered dict, and doing compare
operations to it during initial load.

The ordered dict loading is far more expensive than straight stream
loading, so a cache mechanism for file => loaded yaml is also
added. This reduces the sphinx build time to less than it was
previously, even though we are doing more processing of the data.

Once the parameters.yaml file is made compliant, we can move these
from .info to .warn and enforce the file stays alphabetized.

Part of bp:api-ref-in-rst

Change-Id: Id30464d6326aec3ed7d82c178261d686e2509d33