Introduce a CONF flag to determine glance client version
This patch introduces a CONF flag to determine the glance client
version in use. This flag is set to True currently to ensure
that the glance v1 path is honored as default. However, once the
nova glance proxy is updated to use glance v2 APIs and we are ready
to discard the glance v1 APIs, this flag shall be removed
completely.

Related to blueprint use-glance-v2-api

Change-Id: I90396d71aeac710a13746a10d0218d06dd7503c0