Split the glance API path based on config
In order to get rid of the glance V1 APIs, we need to distinctly
segregate the working paths of the V1 and V2 APIs. This would
ensure smoother transition to V2 than putting in multiple condition
checks in the code.

This approach could have been done using a factory pattern but that
involves greater code change. Since we have a config boolean flag
at our disposal, we use it to split out the code between V1 and V2.
The GlanceImageService doesn't change and acts as V1 while a new
V2 class is written without any implementation at the moment.
Subsequently, all the CRUD operations would be replicated inside
the V2 path. This may lead to duplication of code but having a base
class might involve further re-factoring and building pretty code
that won't be useful in the future. Once we are completely
sure about moving to V2, we should be able to delete the code path
related to V1 and in turn avoid the duplication.

partially implements bp use-glance-v2-api

Co-authored-by: Mike Fedosin <mfedosin@mirantis.com>

Change-Id: I8fbb36530a2763aa779d51d8b027a00913f99afc