Config options: improve libvirt help text (2)
This commit adds additional help text to the following
subset of options: live_migration* and snapshot_image_format.

This commit also formats the help text of some of the other config
options and removes parts of the help texts of other config options
we agreed on to clean up.

Change-Id: I92ffe5022a6b52eb66e134d8acc61684fa0f893e
Implements: blueprint centralize-config-options-ocata