Fall back to raw config drive if not found in rbd
Commit adecf780d3ed4315e4ce305cb1821d493650494b added support for
storing config drives in rbd. Existing instances however still
have config drives in the instance directory. If an existing
instance is stopped, an attempt to start it again fails because the
guest config is generated assuming a config drive location in rbd.

This adds a fall back to the instance directory in the case of
config drive and rbd if the image is not found in rbd.

Conflicts:
	nova/tests/unit/virt/libvirt/test_driver.py
    nova/virt/libvirt/driver.py

    (Changed image_type from 'flat' to 'raw' and image.exists()
    to image.check_image_exists(), removed unit tests that
    don't exist on stable/liberty)

Closes-Bug: #1582684

Change-Id: I21107ea0a148b66bee81e57cdce08e3006a60aee
(cherry picked from commit f5c9ebd56075f8eb04f9f0e683f85bacdcd68c38)