Fall back to flat config drive if not found in rbd
Commit adecf780d3ed4315e4ce305cb1821d493650494b added support for
storing config drives in rbd. Existing instances however still
have config drives in the instance directory. If an existing
instance is stopped, an attempt to start it again fails because the
guest config is generated assuming a config drive location in rbd.

This adds a fall back to the instance directory in the case of
config drive and rbd if the image is not found in rbd.

Closes-Bug: #1582684

Change-Id: I21107ea0a148b66bee81e57cdce08e3006a60aee