deprecate config option "fatal_exception_format_errors"
The config option "fatal_exception_format_errors" got introduced purely
for testing purposes with commit 2e80b297. This isn't an option the ops
would change from it's default value.
This change deprecates the option. When this option gets removed in the
future, we can hardcode that wrong formats should be fatal by default,
as it is already set today in "nova/conf_fixture.py".

Change-Id: I184b1b1cbd85f136cee120965e38e2ef87bb66ff