objects: new InstanceDeviceMetadata object
The InstanceDeviceMetadata will keep a list
of all the devices metadata, such as NICs and disks
associated to an instance. Eventually, it will be
stored in the instance_extra table.
InstanceDeviceMetadata object is new in this series
and has never been sent over RPC nor used.

Partially implements blueprint virt-device-role-tagging
Co-authored-by: Artom Lifshitz <alifshit@redhat.com>

Change-Id: I914e2c9a465864e5d320f8eb8d2fd509e64a7a75