be more explicit that rate limits are gone in v2.1
The rate limit code was trivially wrong the moment you had > 1 API
server, so we deprecated it with the legacy v2 code. The key still
exists in v2.1 but it's always empty. It's just confusing that it
looks like it's still getting data from the environment, which it is
not.

Clear this up and leave some comments so we can further eliminate
things after the v2 legacy drop.

Because there is a difference between v2 legacy and v2.1, delete all
the v2 legacy tests. This included a ton of direct testing of the
Ratelimit middleware that is about to be removed (as it only exists
for the v2 legacy stack).

Change-Id: Icfd77aecf77ad626c0efb93f59a6db11f21b3d56