Remove tests for missing get_available_resource()
There are no virt drivers that do not return a dict of resources
information in the virt driver get_available_resource() call. This means
the unit tests that validate that the resource tracker doesn't do
anything if nothing is returned from the get_available_resource() are
pointless.

We remove the code in the resource tracker that tests to see if None is
returned by the virt driver's get_available_resource() call because no
virt driver actually does this. It's just bloat code.

Change-Id: I66739c05d894a0d6234eda3bdf450acc49e99166