rt-unit: Replace hard-coded strings with constants
The test_tracker.py file over time had had the host and
hypervisor_hostname (nodename) values hard-coded as strings throughout
the file. This commit simply replaces those strings with references to
module-level constants in order to simplify and make the test cases more
consistent. Some of the test cases were using the string 'fake-host' to
refer to the host and 'fakehost' to refer to the nodename. Other tests
were using the string 'fake-node' to refer to the nodename. This fixes
those all up to be consistent and refer to two constants.

Change-Id: I4158f5661cf083e224171ff47103f77a79a94125