Add ConsoleAuthToken object
Console connection authorizations will be saved in
the database instead of in memory in a console auth
server.

The ConsoleAuthToken object has class methods
that replace the rpc methods used by the console
auth service. To use this object instead of that
service the calling code will use these methods
instead of call the service rpc API.

Co-Authored-By: Eli Qiao <qiaoliyong@gmail.com>
Co-Authored-By: melanie witt <melwittt@gmail.com>

partially-implements: blueprint convert-consoles-to-objects

Change-Id: Ia13591906ab2b3b7b7d5bc8f4b965c117f08fc9b