Remove unused WsgiLimiter
There is unused class WsgiLimiter. The class exists in legacy v2 API
code, and the class will be removed with whole files. However, we need
to remove the class separatelly because it uses DEFAULT_LIMITS which is
necessary to be moved into newer v2.1 API code to operate unit tests.
So this patch removes WsgiLimiter.

Change-Id: I7234217991e83023894e2c2a2e9303a4186ff8dd