Ironic nodes with instance_uuid are not available
Currently, if a node is in AVAILABLE or NOSTATE (legacy), regardless of
if it has an instance_uuid it is considered able to be scheduled to.
However, it's impossible for a deployment to succeed to an ironic node
with instance_uuid populated. We should not schedule to nodes in this
state.

Change-Id: I41e0c8f1f8a91e11180a6edd72907cf76fe4b235
Closes-bug: 1503453