Raise exception if BuildRequest deleted twice
After an instance has been scheduled to a cell and mapped to the
instance table there the BuildRequest for it will be destroyed.  An
instance delete that occurs before that remapping has happened will also
destroy the BuildRequest.

In order to know that a delete has happened, and stop the build, the
destroy method should raise an exception if it is already deleted.

Change-Id: Ib86ac14e2bd933bb9e76ea86b2a1ec4f40d204a3