Remove oversubscribe test in test_resource_tracker
Oversubscription -- i.e. taking the xxx_allocation_ratio fields into
account when calculating whether a compute node can host an instance --
is something that is handled in the scheduler. There was a unit test in
test_resource_tracker.py that was trying to test something about
oversubscription but was actually doing nothing more than testing the
same code paths that unit tests in test_tracker.TestInstanceClaim are
already doing: that the math of consuming or freeing the claimed amount
of requested resources was correct. So, let's remove this duplicate and
misleading unit test.

Change-Id: If462b150c24411fc7cc7a677fbe3467e95f8d081