Remove pointless "additive claims" unit test
There was a test in test_resource_tracker.py that was, no joke,
repeating an instance claim of the same flavor and checking that the
amount used from a previously-empty node was 2 times the resource
amounts associated with the flavor. This is a useless test and doesn't
exercise any more code paths that existing unit tests in
test_tracker.TestInstanceClaim already stress.

Change-Id: I4b8d4c701ad7db1469a55835067377a58c12884c