Add proxy middleware to application pipeline
Recently the http_proxy_to_wsgi middleware was added before the
versions resource; This correctly sets the URL scheme for keystone
version discovery, however, this is missing the URL schemes for
pagination URLs. So for this to work, the middleware needs to be
added to the application pipelines. Note that a similar commit
was done in cinder [1].

[1] If5aab9cc25a2e7c66a0bb13b5f7488a667b30309

Change-Id: I808469f24066d382decf55b9dad5312d6e068da7
Closes-Bug: #1573766