Update functional tests for fixtures 3
In order to use the 3.0.0 release of the fixtures library some tests
needed to be updated in order to be compatible with 1.4 and 3.0.

Most failures with 3.0 were due to using bound methods of a class to
patch another class. This requires a different signature on the method
used for patching between the fixtures versions. Moving the methods to
be functions addresses this for the cases where the test method does not
need to access the class it was defined on.

For test_networks a different tact was needed. Because the methods used
for patching require accessing state on their defined class a proxy
method is introduced. This proxy method captures and discards the 'self'
parameter from the call point of the patched method. By doing so the
patching methods can have the same signature for multiple fixtures
versions.

Change-Id: I9c5635d0930b091e1ae0a1439d687dd20a0ff939