Expand initial feature classification lists
Lets focus on server operations that must be tested for each of the Nova
virt drivers, such as create a server and snapshot a server.

This initial list is not meant to be complete, rather its some of the
key headline server actions that we should look at first. This list
includes all actions of the above type that are included in DefCore.

Follow changes will look at correctly populating the data for each of
the features that have been included.

blueprint feature-classification

Change-Id: Ibe31acbae38ea9f2b06aedded66aa486ed1f5cb8