Fix nova-compute start failed when reserved_huge_pages has value.
The problem is due to this change:
https://review.openstack.org/#/c/292499/20/nova/conf/virt.py

The code of item_type=types.Dict is not correct, modfiy by
item_type=types.Dict().

Modify the usage discription of config of reserved_huge_pages.
Usage of oslo_config.types.Dict is key:value pairs separated by
commas.

Add unit test because this is a complicated option that might not work
like people expect.

Closes-Bug:#1588305
Change-Id: I06490866f24617cf99764ede73a1938c2d7b7b5c