Remove "or 'reserved'" from _create_volume_bdm
Just like in change I5cde374a2621b22b094d14e357a399a2fb784287,
the "or 'reserved'" logic when creating the BDM when there
isn't a volume_id is vestigial - a carry over from grizzly
and compute RPC API 2.0 which we don't use anymore.

So let's remove this since it's not used and confusing.

Change-Id: Ib06126c5c878bca66708dbe4ad95df469cbe6710
Related-Bug: #1533871