Add periodic task to clean expired console tokens
Console auth tokens and related connection information will be stored
in the database. Each compute manager will be responsible for cleaning
up expired tokens related to instances it manages as well as cleaning
up when instances are deleted or moved.

This patch adds a periodic task to the compute manager to cleanup
expired tokens. It also adds token cleanup for instances when they are
deleted or live migrated. TODOs have been added to add token cleanup
for other move operations such as resize, evacuate, and shelve offload.
They are not changed as part of this patch series because for now, we
are keeping parity with existing behavior. Changing the behavior will
be a separate future patch.

Later patches in this series will add the code to put console auth
tokens in the database.

The consoleauth server will be deprecated. During the deprecation
period tokens will be stored in both the consoleauth server and the
databaes but console proxies will chose which to use for authorization.

partially-implements: blueprint convert-consoles-to-objects

Change-Id: I66762703709340a2f5c68dcd6802993c9a68c263