Use ConsoleAuthToken object to generate authorizations
The get_<type>_console() functions in the compute manager
get a console connection from the hypervisor and generate
an authorization token and an authorized access url.
These are eventually passed to the consoleauth service
to be stored in memory.

This patch changes those functions so they use the
ConsoleAuthToken object. Token generation is encapsulated
in the ConsoleAuthToken.authorize() function that also
saves the authorization in the database. Access url
generation is encapsulated by the object's access_url
property.

A following patch will add a switch to allow the proxies
to use the database for authorization instead of the
consoleauth service. The consoleauth service will be
deprecated when there are no more components that rely
on it.

partially-implements: blueprint convert-consoles-to-objects

Change-Id: I6cc46b3347a426fd27c705abf727801ed47314e4