Cells: Handle delete with BuildRequest
When an instance delete is issued the instance could be in a few
different states which require different handling.

If the instance is in a cell and booted or booting then the delete can
proceed normally.

If there is a BuildRequest object for the instance then it is still
being built and has not reached a cell yet. In this situation there may
or may not also be an instance record depending on how much future work
has merged. The BuildRequest should be destroyed, any InstanceMapping
that exists should be cleaned up, and if there is an Instance it should
be deleted normally.

The added functional test does not check for a change in behavior. It
ensures that the behavior of delete does not change as the delete
process is updated. Note that the only delete path that bails early is
not yet reachable. The test keeps things honest as that changes.

Change-Id: I27d89143e8804aebdd414791f03b622abdaafe90
Partially-implements: bp add-buildrequest-obj
Partially-implements: bp cells-scheduling-interaction