Create Instance from BuildRequest if not in a cell
This is preliminary work towards moving instance creation until after
scheduling. In order to do that we still need the ability to honor an
instance show request. This does a lookup to determine if an instance is
in a cell and if not it synthesizes an instance from the BuildRequest
object that has been stored.

If the instance is not mapped then the deployment has not migrated to
their first cell yet and the instance lookup proceeds as usual.

Until instance creation is moved to after scheduling this is mostly a
noop.

Honoring the instance list command will come later.

Change-Id: Ib984c30543acb3ca9cb95fb53d44d9ded0f5a5c8
Partially-implements: bp add-buildrequest-obj
Partially-implements: bp cells-scheduling-interaction