Only chown console log in rescue
Change I38a97df92330a9f20f43e9e4598da3375c6a22dc introduced a hack to
chown console.log to fix an issue during rescue. For some reason this
hack was added to _create_image. As it is not related to image
creation and (according to the bug) is only relevant to the rescue
path, we move it to rescue.

Change-Id: I35c55a284681533b344c965f67a2c8f3f99f1c37