Add set_inventory() method on ResourceProvider
Adds a set_inventory() method on the ResourceProvider object that
replaces the resource provider's inventory with a supplied set of
inventory records in an InventoryList object. The replacement is done in
a consistent fashion using the resource provider's generation value to
ensure that a writer doesn't update the inventory of a resource provider
after another writer has changed the inventory in between the original
read of inventory information and the subsequent save of that inventory.

A functional test is in place which attempts to exercise all three
of add, delete and update inventory as well as causing a
ConcurrentUpdateException.

The object hash has changed but no version increment done because
the ResourceProvider object is not yet in use in the wild. These
changes are still bringing the eventual first version to its correct
form.

Partially-Implements: blueprint generic-resource-pools
Change-Id: I2214fa5d176269e65fdcd7a5a7e4958a925235d3