Remove manual creation of console.log
libvirt will automatically create a configured console log when it
starts a guest. However, the libvirt driver manually creates
console.log in 2 places: _create_image and pre_live_migration.

console.log creation was added to _create_image in commit 58e18861,
which was before gerrit. The reason in the commit message is:

  Create console.log ahead of time. This ensures that the user running
  nova-compute maintains read privileges.

This was later patched to fixup permissions if _create_image runs
again, because libvirt chowns it when it runs. Additionally, code has
been added at the point of use, get_console_output, to chown it
immediately before trying to read it. I believe this last case is the
only relevant one here. As this seems to obsolete the original patch,
we remove it.

The change to pre_live_migration came in change If7abd63a. The
specific reason for adding it is not given in the commit message, but
there is a code comment: 'Touch the console.log file, required by
libvirt'. I don't believe this is the case, so we remove it.

Change-Id: Ia84579b84105e8ad6146591188a96e847e762c77