Remove duplicate test of DELETED instances
test_resource_tracker had a test to ensure that DELETED instances are
not accounted for in the audit process that runs in the RT's
update_available_resource() method. test_tracker's instance fixtures
already included a DELETED instance and we were correctly checking in
a unit test that the DELETED instance's used resources were not
accounted for by the RT. Added a note in the test_tracker unit test
about this and deleted the duplicate unit test from
test_resource_tracker.py.

Change-Id: I31b29244b77032af19b0790b2c5d02037b6c332a