Remove max_size argument to images.fetch and fetch_to_raw
images.fetch was passed a max_size argument, but did not use it.

images.fetch_to_raw used the max_size argument to check that the
image being downloaded is not larger than target instance's root disk.
However, this check does not make sense in this context. fetch_to_raw
is used to download directly to the image cache, which means when
booting multiple instances on the same compute it only executes once.
However, the check obviously needs to happen against every instance,
not just the first to use a particular image. Consequently every image
backend duplicates this check, making the check in fetch_to_raw both
confusing and redundant.

There are a couple of callers outside the libvirt driver. These do not
pass the max_size argument, and are therefore unaffected.

Implements: blueprint libvirt-instance-storage

Change-Id: I70a559f3dc9b59097ff6923920f4377cca00d1b2