libvirt: split off code for updating live migration downtime
Periodically the migration code needs to configure an
updated downtime with the QEMU guest. Move the code for
doing this out into its own method to make it possible
to test it directly. This changes the way we check the
downtime - instead of incrementally popping entries
off the downtime_steps list, we leave the list intact
and simply keep track of the current downtime value.
This avoids having side-effects on the parameters passed
into the downtime update method.

Implements: blueprint libvirt-clean-driver
Change-Id: I2e98abb519677f442cd79ea7fa63209bf11a0307