Change default fake_ server status to ACTIVE
Extensions like the hide address extension operate on the server
status. In the before time, we picked BUILD as the default state of a
fake server. Most of the time we actually want ACTIVE fake servers for
tests. Changing this default to ACTIVE makes some of the follow up
patches around extensions to require less changes because we don't
need to change a whole bunch of tests to individually set the status
to ACTIVE.

Change-Id: I4ed13a6a2f466a9dbf3fc32bea46c456ab2e6c40