libvirt: split out code for processing migration tasks
While migration jobs are running it is possible to have
tasks requested against the VM, such as pausing it. This
task processing is handled by a separate inline method.
Move this inline method out to be a static method in
the class, and add explicit testing of it. This also
adds an explicit warning if it finds an unknown task.

Implements: blueprint libvirt-clean-driver
Change-Id: Ica10a704b92c3fabe80f9f2aac54d947f0c3d147