nova-net: handle 'auto' network request in allocate_for_instance
This adds the check to the nova-network API allocate_for_instance
method such that if no network is allocated for the instance and
'auto' was requested, we raise an error.

Also, if 'auto' is requested, we pass requested_networks=None
to the network rpc api (and network manager) since a network
with uuid 'auto' isn't going to exist.

Similar handling is done for deallocate_for_instance.

Some of the unit tests for allocate had to be fixed to pass
a valid requested_networks value.

Part of blueprint get-me-a-network

Change-Id: I718e2b3312ddd9fab0808a585bdbefbb5419f901