libvirt: Fix fake _disk_info data in LibvirtDriverTestCase
This fake was returning dicts which were missing backing_file and
over_committed_disk_size. This change adds the missing fields with their
correct default values. We also move it out of LibvirtDriverTestCase
to the top level so tests in other suites can use it in future
patches.

The data returned was also not related to the instance being tested.
This meant that _test_finish_migration was explicitly testing for the
wrong behaviour, as it was expecting a swap disk to be created for an
instance which did not have one.

We also update LibvirtDriverTestCase._create_instance to use
fake_instance.fake_instance_obj, which helps avoid some accidental DB
access in tests.

Change-Id: If2ebba557b05330387141647049cf9a5cbe6aa34