Add delete_, update_ and add_ inventory to ResourceProvider
Each of these reuses the methods used by set_inventory for
deleting, updating or adding just one Inventory to a
ResourceProvider.

With the addition of these methods, the existing set_inventory
needs a clarifying docstring.

An Inventory.capacity property has been added to validate that
Inventory.total and Inventory.reserved do not conflict. This will
also eventually be used to make comparisons with Usages.

The version of the ResourceProvider object does not require a
version bump, despite the change in object hash, because
ResourceProviders are not yet being used in the wild.

Partially Implements: blueprint generic-resource-pools
Change-Id: I62836faf829b93997a176aba4e0dc3502ae54049