New discover command to add new hosts to a cell
While there are some nova-manage commands to take an existing deployment
and migrate all of its hosts into a new cellsv2 environment there was no
way to add more hosts to a cell. This command can be run at any time
after the initial migration and will map any hosts in a cell that have
not been seen before.

Nothing else changes about adding hosts to a deployment. Configure them
to use a nova database and start them up and they'll register themselves
with that database. This new command simply lets the API know how to
route requests to those hosts. Until this is done instances can not be
booted on those hosts.

Partially-Implements: bp cells-scheduling-interaction
Change-Id: I8c044e5b480edddead28d8c3527d003da566ed1e