Add a ResourceProviderList object
The Placement API needs a way to list all the resource providers.
Currently this list is made up of all resource providers for which
can_host == 0. This is to avoid have a list of a few thousand
compute nodes alongside the stuff we want to be able to update
in a manual fashion.

A functional test confirms that created resource providers are in
fact listed when we get_all.

Co-Authored-By: Sylvain Bauza <sbauza@redhat.com>

Partially Implements: blueprint generic-resource-pools
Change-Id: Ie1f76d17baa758e379afc8aed42c96c668759581